Rename the popup charts component.
