+++ meteor-app/imports/ui/components/charts-popup.js
